backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
python-dotenv>=1.0.1
httpx>=0.28.0
google-genai>=1.5.0
elevenlabs>=1.50.0
python-multipart>=0.0.6
fal-client>=0.13.0
